// src/product-phone/dto/create-product-phone.dto.ts
/* eslint-disable prettier/prettier */
export class CreateProductPhoneDto {
    id?: string;
    name?: string;
    cash?: number;
    display?: string;
    inch?: string;
    ram?: string;
    memory?: string;
    img?: string;
    operating_system?: string;
    installment?: number;
    type: string;
}